# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "My API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "Please enter a valid token to test the requests below..."
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/users/me": {
      "patch": {
        "tags": [
          "User"
        ],
        "description": "",
        "parameters": [
          {
            "name": "edit-user",
            "in": "body",
            "schema": {
              "required": [
                "name"
              ],
              "properties": {
                "name": {
                  "type": "string",
                  "maxLength": 250,
                  "example": "thang mai"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "get": {
        "tags": [
          "User"
        ],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/users/change-password": {
      "patch": {
        "tags": [
          "User"
        ],
        "description": "Endpoint change password .",
        "parameters": [
          {
            "name": "change-password",
            "in": "body",
            "schema": {
              "required": [
                "password",
                "oldPassword"
              ],
              "properties": {
                "password": {
                  "type": "string",
                  "maxLength": 250,
                  "minLength": 6,
                  "example": "123456"
                },
                "oldPassword": {
                  "type": "string",
                  "maxLength": 250,
                  "minLength": 6,
                  "example": "123456"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/users/upload": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "Endpoint change image avatar .",
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "singleFile",
            "in": "formData",
            "type": "file",
            "extensions": ".png|.jpeg.|jpg",
            "required": true,
            "description": "choose image with extension is supported."
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    }
  },
...
}